Create ipChecker.py
This program solves the problem I run into a lot. The IP Address of the Raspberry Pi changes, and without knowing what it changes too, it makes it hard to remote into the device.

I'll update this program to eventually note changes in the Public IP Address, to help for when that changes (if you need to SSH into the device).